+++ pi-monitor/pi-monitor-client/sdks/full/src/commonMain/kotlin/pimonitor/evaluation/business/forms/CreateBusinessIntent.kt
@@ -1,0 +1,5 @@
+package pimonitor.evaluation.business.forms
+
+sealed class CreateBusinessIntent {
+    data class ShowForm(val inviteId: String?) : CreateBusinessIntent()
+}