+++ pi-monitor/pi-monitor-core/src/commonMain/kotlin/pimonitor/authentication/signup/SignUpParamsUtils.kt
@@ -1,4 +1,4 @@
-package pimonitor.monitors
+package pimonitor.authentication.signup
 
 import bitframe.authentication.signin.SignInCredentials
 import bitframe.authentication.spaces.RegisterSpaceParams