+++ pi-monitor/pi-monitor-core/src/commonMain/kotlin/pimonitor/monitored/CreateMonitoredBusinessParams.kt
@@ -1,0 +1,7 @@
+package pimonitor.monitored
+
+data class CreateMonitoredBusinessParams(
+    val businessName: String,
+    val contactName: String,
+    val contactEmail: String
+)