+++ pi-monitor/pi-monitor-services/ktor/src/commonMain/kotlin/pimonitor/authentication/signup/SignUpServiceKtor.kt
@@ -11,7 +11,6 @@ import io.ktor.client.statement.*
 import kotlinx.serialization.json.Json
 import later.Later
 import later.later
-import pimonitor.monitors.SignUpParams
 
 class SignUpServiceKtor(
     override val bus: EventBus,