+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## PiMonitor Client SDK Core
 
 - [[PM124]](https://github.com/picortex/bitframe/issues/124) Removed onUserLoggedIn method from SignInViewModel
+- [[PM141]](https://github.com/picortex/bitframe/issues/141) Created a Register business form 
 
 # 0.0.30